fix: set fall interval on each fall stopwatch tick
Previously (by developer error) the interval for the fall stopwatch was
only being updated when the function `ToggleSoftDrop` was called. Now,
`ToggleSoftDrop` is only responsible for doing the toggle not also
returning the new value. A new function `GetFallInterval` has been
created to retrieve the fall interval. This allows the toggle to be done
on soft drop key press, and the stopwatch interval is only modified when
the stopwatch tick message is received.